requirements.txt
git+https://github.com/openai/whisper.git
moviepy
pillow
